downgrade avx version to avx2

+++ simd_benchmarking/image.c
@@ -6,9 +6,9 @@ typedef struct {
     int32_t min;
 } MaxMin;
 
-MaxMin max_min_in_list(int32_t* list, int32_t size) {
-    int32_t max = list[0];
-    int32_t min = list[0];
+MaxMin max_min_in_list(float* list, int32_t size) {
+    float max = list[0];
+    float min = list[0];
 
     for (int32_t i = 1; i < size; i++) {
         if (list[i] > max) {
@@ -32,11 +32,11 @@ void write_image(Matrix image, const char *filename) {
   fprintf(file, "P5\n %s\n %d\n %d\n %d\n", comment, image.width, image.height, 255);
 
   MaxMin max_min = max_min_in_list(image.data, image.width * image.height);
-  int32_t max = max_min.max;
-  int32_t min = max_min.min;
+  float max = max_min.max;
+  float min = max_min.min;
 
   for (int i = 0; i < image.width * image.height; i++) {
-      int32_t data = image.data[i];
+      float data = image.data[i];
       uint8_t pixel = (data - min) * 255 / (max - min);
       fputc(pixel, file);
   };
@@ -50,7 +50,7 @@ Matrix read_image(const char *filename) {
     int width, height;
     fscanf(file, "P5\n%d %d\n255\n", &width, &height);
 
-    int32_t *data = malloc(width * height * sizeof(int32_t));
+    float *data = malloc(width * height * sizeof(float));
 
     for (int i = 0; i < width * height; i++) {
         data[i] = fgetc(file);
@@ -68,7 +68,7 @@ Matrix read_image(const char *filename) {
 void print_image(Matrix image) {
     for (int i = 0; i < image.height; i++) {
         for (int j = 0; j < image.width; j++) {
-            printf("%06d ", image.data[i * image.width + j]);
+            printf("%06d ", (int) image.data[i * image.width + j]);
         }
         printf("\n");
     }
